Rent My Kids NFT
This Members Only Utility is  an earned bonus for holders. When Users have 100 KIDS token, they can redeem it for "Rentable" NFT. These will be Minted in batches, and then airdropped to the respective address.  This will avoid KIDS from having to pay any gas fees. The NFT are an ERC4907, which are able to differentiate user from holder, giving KIDS a whole new world of possibilities. They also unlock perks within the community as many aspects are token gated, or level up.